+++ internal/dhcpd/os_windows.go
@@ -1,13 +1,17 @@
+// +build windows
+
+//go:build windows
+
 package dhcpd
 
 import (
 	"net"
 
-	"github.com/AdguardTeam/golibs/errors"
+	"github.com/AdguardTeam/AdGuardHome/internal/aghos"
 	"golang.org/x/net/ipv4"
 )
 
 // Create a socket for receiving broadcast packets
-func newBroadcastPacketConn(bindAddr net.IP, port int, ifname string) (*ipv4.PacketConn, error) {
-	return nil, errors.Error("newBroadcastPacketConn(): not supported on Windows")
+func newBroadcastPacketConn(_ net.IP, _ int, _ string) (*ipv4.PacketConn, error) {
+	return nil, aghos.Unsupported("newBroadcastPacketConn")
 }